tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-fill-mode_forwards_removing_animation.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-fill-mode: forwards, removing
   animation</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-fill-mode-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="If 'animation-fill-mode' is 'forwards', it
   applies until the animation style is removed">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            margin-left: 0px;
         }
         to {
            margin-left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         animation-duration: 0.5s;
         animation-fill-mode: forwards;
      }
      .test {
         animation-name: slide;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(div, null);
         setTimeout(function() {
            var passed = (style.marginLeft == '100px');
            div.className = '';
            passed = (passed && style.marginLeft == '0px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         }, 1000);
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should slide to the right and then pause for a while
   before jumping back to its initial position.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
